Added a module that allows for translation of German words into English, using the public API of https://glosbe,.com .
Note that especially GlosbeModule.cs was NOT error checked as I didn't have access to the original libraries used when creating the project and VS wouldn't work when I tried downloading and linking those by hand. To prevent errors even in this case, I oriented as much as possible towards the style of JishoModule.cs, essentially copying it and changing the API-dependent bits.